requirements.txt
# Core dependencies
python>=3.8
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
ftfy>=6.1.0

# Web scraping & data collection
selenium>=4.10.0
webdriver-manager>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tweepy>=4.14.0

# Data processing
polars>=0.18.0  # Alternative for pandas for better performance
joblib>=1.3.0
tqdm>=4.65.0

# Visualization
plotly>=5.15.0
dash>=2.11.0
streamlit>=1.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Natural Language Processing
transformers>=4.30.0
sentence-transformers>=2.2.0
indic-nlp-library>=0.80  # For Indian language processing

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
click>=8.1.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Async/Concurrency
aiohttp>=3.8.0
asyncio>=3.4.3
concurrent-futures>=3.1.1

# System monitoring
psutil>=5.9.0
memory-profiler>=0.61.0
